Begun fleshing out openGL implementation
