+++ Include/UI/geometryEditor2D.h
@@ -4,7 +4,11 @@
 /*
 	This file will contain all of the definitions for the 2D Geometry Editor
 */
-
+#include <wx/wx.h>
+#include <wx/glcanvas.h>
+#include <wx/bmpbuttn.h>
+#include <wx/window.h>
+#include <UI/common.h>
 
 /*! \class geometryEditor2DAbstraction
 	\brief The main class for the 2D geometry Editor
@@ -27,7 +31,7 @@ public:
 	*************/
 private:
 	
-}
+};
 
 
 
@@ -53,7 +57,7 @@ public:
 private:
 	
 //! The object for the abstract layer of the 
-}
+};
 
 
 
@@ -62,14 +66,17 @@ private:
 	
 	This is the presentation part of the 2D geometry editor
 */
-class geometryEditor2DPresentation
+class geometryEditor2DPresentation : public wxPanel
 {
 	/**********************
 	* Function prototypes *
 	***********************/
 public:
 	//! The constructor for the class. This is what will get called first in order to create and manipulate 2D geometry
-	geometryEditor2DPresentation();
+	geometryEditor2DPresentation(wxWindow *par, const wxPoint &position, const wxSize &size);
+	void createOpenGLCanvas();
+	
+private:
 	
 	
 	
@@ -77,8 +84,10 @@ public:
 	* Variables *
 	*************/
 private:
-	
-}
+	wxGLCanvas *drawingCanvas;
+	wxGLContext *drawingContext;
+//	wxGLAttributes dispAttributes;
+};
